Refactor phase 2b: split image_export.py into a package
Continues the effort tracked in docs/development/LARGE_FILE_REFACTOR_PLAN.md.

app_utils/image_export.py (3391 lines) became the app_utils/image_export/
package. The share-image renderer mixed thirteen concerns in one file:
brand logo, canvas layouts, colour palette, fonts, the ALL-CAPS
humanizer, threat icons, event theming, drawing primitives, weather
particle effects, OSM tile fetching, map rendering, info-panel drawers
and the top-level composer. Each is now its own module, with an acyclic
dependency graph generated from the actual imports. The package
__init__.py re-exports all 125 names the single-file module exposed, so
app_core/notifications/alert_image.py and webapp/admin/api.py are
unchanged.

Verified as pure motion by comparing ast.dump() of every top-level
definition: 68 definitions, 68 matches, zero differences. The slicing
step also asserted every non-blank line of the original landed in
exactly one module, so nothing was silently dropped.

Fix a real bug the split introduced. _LOGO_PATH and
_TILE_DISK_CACHE_DIR_DEFAULT walk two directories up from __file__ to
reach the repository root - correct for a single app_utils/image_export.py,
one level short for a module inside the package. The share card would
have rendered with no brand logo (_load_logo swallows the error) and
tiles would have cached into app_utils/data/tile-cache, colliding with
the directory 2.134.0 created for the FIPS table. Neither failure raises
and all 120 existing tests still passed; it surfaced only as a stray
directory in git status. This is the one way a verbatim move can change
behaviour - the code really is identical, so the AST check cannot see
it - so the plan's ground rules gained a rule to audit __file__-relative
paths before any split. Both constants now derive from a named
_REPO_ROOT, and three new tests pin the resolved paths and assert the
logo actually loads; all three fail without the fix.

tests/test_image_export_themes.py needed two mechanical updates, both
because a package has more than one namespace where the module had one.
It loads the renderer by file path to avoid importing all of app_utils,
which now needs submodule_search_locations on the spec. And nine
monkeypatch.setattr calls moved to the module that calls the patched
name (tiles for _http, maps for _fetch_tile/_fetch_county_outlines,
render for _render_map) - rebinding on the re-exporting package does not
change what maps._render_map sees in its own globals. Confirmed
load-bearing: pointing them back at the package fails 5 tests.

Also convert app_core/radio/demod/ to relative intra-package imports,
matching webapp/audio_archive/, app_core/flask/, app_core/config/ and
the new image_export/. Beyond consistency, this is what lets a package
be loaded standalone by file path in tests.

Tests: full suite 1973 passed, 67 xfailed, 0 failures.
